Create, install, load and attach multiple disposable packages

If a package with the same name as a disposable one, is loaded, then it will be unloaded. If a package with same name as a disposable on is installed in lib_dir, then it will be overwritten. (lib_dir is usually a temporary directory, so this is not a big problem.)

make_packages(..., lib_dir = tempfile(), imports = character(),
  quiet = TRUE)

Arguments

...

Named expressions. A separate package with the given name is created for each.

lib_dir

Directory to install the package to. Defaults to a temporary directory that is deleted once the R session is over.

imports

The 'Imports' field in the DESCRIPTION file, the packages to import in each disposable package. It can be a character vector, which applies to all packages, or a list of character vectors, one for each package.

quiet

Whether to show the installation process of disposable packages.

Value

A named list with entries:

lib_dir The directory in which the packages are installed.

package The named of the packages.

Details

Note that if you specify lib_dir and it points to an existing directory, make_package overwrites the packages there. If an error happens during installation or loading of the disposables packages, then it will not restore the original contents of lib_dir, but it will remove all newly installed disposable packages, even the ones that were installed cleanly.

Examples

pkg <- make_packages(
  foo1 = { f <- function() print("hello!") ; d <- 1:10 },
  foo2 = { f <- function() print("hello again!") ; d <- 11:20 }
)
foo1::f()
foo2::f()
foo1::d
foo2::d
dispose_packages(pkg)

See also

Examples

## See above and also in the README: https://r-pkg.org/pkg/disposables